chore: move LauncherVersion to `model` pkg
